# Conflicts:
#	backend/src/main/java/com/noblesi/travelplanner/admin/tour/service/AdminTourSyncService.java

+++ backend/src/main/java/com/noblesi/travelplanner/admin/member/dto/AdminMemberDetailDTO.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class AdminMemberDetailDTO {
 
-	private int memberId;
+	private Long memberId;
 	private String memberName;
 	private String nickName;
 	private String email;
